src/requirements.txt
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.3
pyparsing==3.0.7
python-dotenv==0.19.2
redis==4.1.4
requests==2.26.0
rq==1.10.1
SQLAlchemy==1.4.31
urllib3==1.26.7
waitress==2.0.0
Werkzeug==2.0.2
wrapt==1.13.3